--- packages/hierarchies-react/src/presentation-hierarchies-react/itwinui/TreeNodeRenderer.css
@@ -1,29 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-* See LICENSE.md in the project root for license terms and full copyright notice.
-*--------------------------------------------------------------------------------------------*/
-
-.stateless-tree-node {
-  user-select: none;
-}
-
-.stateless-tree-node .action-buttons {
-  visibility: hidden;
-}
-
-.stateless-tree-node:hover .action-buttons,
-.stateless-tree-node.filtered .action-buttons,
-.stateless-tree-node:focus-within .action-buttons {
-  visibility: visible;
-}
-
-.stateless-tree-node:focus-within {
-  outline: 1px solid var(--iui-color-border-accent);
-  outline-offset: -1px;
-  border-radius: var(--iui-border-radius-1);
-}
-
-.stateless-tree-node-small-spinner {
-  height: var(--iui-size-s);
-  width: var(--iui-size-s);
-}
