+++ packages/hierarchies-react/src/presentation-hierarchies-react/itwinui/TreeRenderer.tsx
@@ -3,21 +3,21 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-import { ComponentPropsWithoutRef, useCallback } from "react";
-import { NodeData, Tree } from "@itwin/itwinui-react";
+import { ComponentPropsWithoutRef } from "react";
+import { Tree } from "@itwin/itwinui-react/bricks";
 import { PresentationTreeNode } from "../TreeNode.js";
 import { SelectionMode, useSelectionHandler } from "../UseSelectionHandler.js";
 import { useTree } from "../UseTree.js";
 import { LocalizationContextProvider } from "./LocalizationContext.js";
+import { TreeLevelRenderer } from "./TreeLevelRenderer.js";
 import { TreeNodeRenderer } from "./TreeNodeRenderer.js";
 
-/** @public */
-type TreeProps = ComponentPropsWithoutRef<typeof Tree<RenderedTreeNode>>;
+/** @alpha */
+export type TreeProps = ComponentPropsWithoutRef<typeof Tree.Root>;
 
-/** @public */
-type TreeNodeRendererProps = ComponentPropsWithoutRef<typeof TreeNodeRenderer>;
-
-/** @public */
+/** @alpha */
+export type TreeNodeRendererProps = ComponentPropsWithoutRef<typeof TreeNodeRenderer>;
+/** @alpha */
 interface TreeRendererOwnProps {
   /** Root nodes of the tree. */
   rootNodes: PresentationTreeNode[];
@@ -25,12 +25,11 @@ interface TreeRendererOwnProps {
   selectionMode?: SelectionMode;
 }
 
-/** @public */
-type TreeRendererProps = Pick<ReturnType<typeof useTree>, "rootNodes" | "expandNode"> &
+/** @alpha */
+type TreeRendererProps = Pick<ReturnType<typeof useTree>, "expandNode"> &
   Partial<Pick<ReturnType<typeof useTree>, "selectNodes" | "isNodeSelected" | "getHierarchyLevelDetails" | "reloadTree">> &
-  Pick<TreeNodeRendererProps, "onFilterClick" | "getIcon" | "getLabel" | "getSublabel" | "filterButtonsVisibility"> &
+  Omit<TreeNodeRendererProps, "node"> &
   TreeRendererOwnProps &
-  Omit<TreeProps, "data" | "nodeRenderer" | "getNode" | "enableVirtualization"> &
   ComponentPropsWithoutRef<typeof LocalizationContextProvider>;
 
 /**
@@ -38,77 +37,32 @@ type TreeRendererProps = Pick<ReturnType<typeof useTree>, "rootNodes" | "expandN
  * are rendered using `TreeNodeRenderer` component from this package.
  *
  * @see https://itwinui.bentley.com/docs/tree
- * @public
+ * @alpha
  */
-export function TreeRenderer({
-  rootNodes,
-  expandNode,
-  selectNodes,
-  isNodeSelected,
-  onFilterClick,
-  getIcon,
-  getLabel,
-  getSublabel,
-  getHierarchyLevelDetails,
-  reloadTree,
-  selectionMode,
-  localizedStrings,
-  size,
-  filterButtonsVisibility,
-  ...treeProps
-}: TreeRendererProps) {
+export function TreeRenderer({ rootNodes, expandNode, localizedStrings, selectNodes, isNodeSelected, selectionMode, ...treeProps }: TreeRendererProps) {
   const { onNodeClick, onNodeKeyDown } = useSelectionHandler({
     rootNodes,
     selectNodes: selectNodes ?? noopSelectNodes,
     selectionMode: selectionMode ?? "single",
   });
-  const nodeRenderer = useCallback<TreeProps["nodeRenderer"]>(
-    (nodeProps) => {
-      return (
-        <TreeNodeRenderer
-          {...nodeProps}
-          filterButtonsVisibility={filterButtonsVisibility}
-          expandNode={expandNode}
-          getHierarchyLevelDetails={getHierarchyLevelDetails}
-          onFilterClick={onFilterClick}
-          onNodeClick={onNodeClick}
-          onNodeKeyDown={onNodeKeyDown}
-          getIcon={getIcon}
-          getLabel={getLabel}
-          getSublabel={getSublabel}
-          reloadTree={reloadTree}
-          size={size}
-        />
-      );
-    },
-    [
-      filterButtonsVisibility,
-      expandNode,
-      getHierarchyLevelDetails,
-      onFilterClick,
-      onNodeClick,
-      onNodeKeyDown,
-      getIcon,
-      getLabel,
-      getSublabel,
-      reloadTree,
-      size,
-    ],
-  );
-
-  const getNode = useCallback<TreeProps["getNode"]>((node) => createRenderedTreeNodeData(node, isNodeSelected ?? noopIsNodeSelected), [isNodeSelected]);
 
   return (
     <LocalizationContextProvider localizedStrings={localizedStrings}>
-      <Tree<RenderedTreeNode> {...treeProps} size={size} data={rootNodes} nodeRenderer={nodeRenderer} getNode={getNode} enableVirtualization={true} />
+      <Tree.Root style={{ height: "100%", width: "100%" }}>
+        <TreeLevelRenderer
+          {...treeProps}
+          nodes={rootNodes}
+          expandNode={expandNode}
+          onNodeClick={onNodeClick}
+          onNodeKeyDown={onNodeKeyDown}
+          isNodeSelected={isNodeSelected}
+        />
+      </Tree.Root>
     </LocalizationContextProvider>
   );
 }
 
 function noopSelectNodes() {}
-function noopIsNodeSelected() {
-  return false;
-}
 
 /**
  * A data structure for a tree node that is rendered using the `TreeRenderer` component.
@@ -118,7 +72,7 @@ function noopIsNodeSelected() {
  * while its children are being loaded. This allows the node renderer to show a placeholder under the parent during
  * the process.
  *
- * @public
+ * @alpha
  */
 export type RenderedTreeNode =
   | PresentationTreeNode
@@ -127,48 +81,3 @@ export type RenderedTreeNode =
       parentNodeId: string | undefined;
       type: "ChildrenPlaceholder";
     };
-
-/**
- * An utility function that creates an `@itwin/itwinui-react` `NodeData` object for the `Tree` component from a
- * `RenderedTreeNode` object.
- *
- * Usage example:
- * ```tsx
- * function MyComponent({ rootNodes, nodeRenderer }: MyComponentProps) {
- *   const getNode = useCallback<TreeProps["getNode"]>((node) => createRenderedTreeNodeData(node, () => false), []);
- *   return <Tree<RenderedTreeNode> data={rootNodes} getNode={getNode} nodeRenderer={nodeRenderer} />;
- * }
- * ```
- *
- * @public
- */
-export function createRenderedTreeNodeData(node: RenderedTreeNode, isNodeSelected: (nodeId: string) => boolean): NodeData<RenderedTreeNode> {
-  if ("type" in node) {
-    return {
-      nodeId: node.id,
-      node,
-      hasSubNodes: false,
-      isExpanded: false,
-      isSelected: false,
-      isDisabled: true,
-    };
-  }
-  return {
-    nodeId: node.id,
-    node,
-    hasSubNodes: node.children === true || node.children.length > 0,
-    subNodes:
-      // returns placeholder node to show as child while children is loading.
-      node.children === true
-        ? [
-            {
-              id: `Loading-${node.id}`,
-              parentNodeId: node.id,
-              type: "ChildrenPlaceholder",
-            },
-          ]
-        : node.children,
-    isExpanded: node.isExpanded,
-    isSelected: isNodeSelected(node.id),
-  };
-}
